refactor: extract generated media from AgentWorkspace
